+++ src/main/java/com/larsvansoest/runelite/clueitems/config/EmoteClueItemsConfigProvider.java
@@ -26,7 +26,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.larsvansoest.runelite.clueitems;
+package com.larsvansoest.runelite.clueitems.config;
 
 import com.larsvansoest.runelite.clueitems.overlay.widget.ItemWidgetContainer;
